place data loaders inside custom hook
